+++ spec/spec_helper.rb
@@ -102,6 +102,10 @@ def action_view_template
   template
 end
 
+def silence_deprecator(&block)
+  Datagrid::Utils.deprecator.silence(&block)
+end
+
 # Requires supporting files with custom matchers and macros, etc,
 # in ./support/ and its subdirectories.
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].sort.each { |f| require f }