+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.0.0
+
+[Changes and migration guide](./version-2)
+
 ## 1.8.3
 
 * Fix rails hooking for version 7.1. [#327](https://github.com/bogdan/datagrid/issues/327)